+++ CHANGES.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   previously `"criterion_"` would not match `net.criterion__weight` as set by
   `net.set_params(criterion__weight=w)`
 - skorch pickle format changed in order to improve CUDA compatibility, if you have pickled models, please re-pickle them to be able to load them in the future
+- `net.criterion_` and its parameters are now moved to target device, previously the user had to make sure that parameters such as class weight are on the compute device
 
 ### Fixed